Rename namespace `Swis\JsonApi\Client\JsonApi` to `Swis\JsonApi\Client\Parsers`
